Dependency updates
* breeze removed (replaced with commons-math)
* Scala 2.12 used
* Joda time removed
* scala-csv now used to serialize to CSV

+++ src/main/scala/eu/semberal/dbstress/model/StatsResults.scala
@@ -3,7 +3,7 @@ package eu.semberal.dbstress.model
 import eu.semberal.dbstress.util.ModelExtensions._
 
 case class StatsResults(count: Long, min: Option[Long], max: Option[Long],
-                        median: Option[Long], mean: Option[Double], stddev: Option[Double])
+                        median: Option[Double], mean: Option[Double], stddev: Option[Double])
 
 object StatsResults {
   def apply(l: List[Long]): StatsResults =
